change floatfield to decimal field

+++ finance/models.py
@@ -5,7 +5,7 @@ from erp.users.models import userprofile
 
 class Budget(models.Model):
     name = models.CharField(max_length=1)
-    total_amount = models.FloatField(blank=True,null=True)
+    total_amount = models.DecimalField(max_digits=12,decimal_places=2,blank=True,null=True)
     department = models.ForeignKey(Department)
     comment = models.TextField(blank=True)
     event_core_status = models.BooleanField(default=False)
@@ -16,7 +16,7 @@ class Item(models.Model):
     budget = models.ForeignKey(Budget)
     department = models.ForeignKey(Department)
     name = models.CharField(max_length=50)
-    quantity=models.FloatField(blank=True,null=True)
+    quantity=models.DecimalField(max_digits=12,decimal_places=2,blank=True,null=True)
     description = models.TextField(blank=True)
     comment = models.TextField(blank=True)
     original_amount = models.FloatField()
